requirements.txt
pymongo
python-dotenv
uvicorn 
fastapi
